% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfRAW.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Fonction de densité d'une loi normale multivariée}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{Une matrice de n colonnes et p lignes}

\item{mean}{Un vecteur contenant les n moyennes de colonnes de x}

\item{varcovM}{La matrice de variance covariance}

\item{Log}{True par défault}
}
\value{
Cette fonction retourne la matrice x et le vecteur des densités
}
\description{
Cette fonction permet de calculer la densité d'une loi normale multivariée
}
